refactor: split reset detector support
